requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
flask-restx==1.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.3.0
PyJWT==2.8.0
pytest==7.4.1
pytz==2023.3.post1
referencing==0.30.2
rpds-py==0.10.2
Werkzeug==2.3.7
